requirements.txt
certifi==2019.9.11
chardet==3.0.4
Django==2.2.5
django-extensions==2.2.1
django-filter==2.2.0
django-rest-framework==0.1.0
djangorestframework==3.10.3
idna==2.8
lazy==1.4
mock==3.0.5
pytz==2019.2
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.5
